+++ resources/graphics.py
@@ -392,6 +392,7 @@
 				"multiplySmall",
 				"replaceSmall",
 				"createSmall",
+				"createIfMissingSmall",
 				"lessThanSmall",
 				"greaterThanSmall",
 				"listAppendSmall",