+++ src/GafferSceneUI/ParameterInspector.cpp
@@ -98,6 +98,12 @@ IECore::ConstObjectPtr ParameterInspector::value( const GafferScene::SceneAlgo::
 	return shader->parametersData()->member( m_parameter.name );
 }
 
+IECore::ConstObjectPtr ParameterInspector::fallbackValue( const GafferScene::SceneAlgo::History *history ) const
+{
+	// No fallback values are provided for parameters. Implemented to override AttributeInspector::fallbackValue().
+	return nullptr;
+}
+
 Gaffer::ValuePlugPtr ParameterInspector::source( const GafferScene::SceneAlgo::History *history, std::string &editWarning ) const
 {
 	auto sceneNode = runTimeCast<SceneNode>( history->scene->node() );