+++ Changes.md
@@ -49,7 +49,16 @@ Breaking Changes
 1.4.x.x (relative to 1.4.9.0)
 =======
 
+Improvements
+------------
 
+- LightEditor : Values of inherited attributes are now displayed in the Light Editor. These are presented as dimmed "fallback" values.
+- LightEditor, RenderPassEditor : Fallback values shown in the history window are displayed with the same dimmed text colour used for fallback values in editor columns.
+
+Fixes
+-----
+
+- LightEditor, RenderPassEditor : Added missing icon representing use of the `CreateIfMissing` tweak mode in the history window.
 
 1.4.9.0 (relative to 1.4.8.0)
 =======
@@ -709,7 +718,10 @@ Build
 1.3.16.x (relative to 1.3.16.6)
 ========
 
+Fixes
+-----
 
+- LightEditor, RenderPassEditor : Added missing icon representing use of the `CreateIfMissing` tweak mode in the history window.
 
 1.3.16.6 (relative to 1.3.16.5)
 ========